requirements.txt
snowflake-connector-python>=3.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
